+++ example/index.js
@@ -1,0 +1,15 @@
+export {default as creator} from "./src/creator";
+export {default as local} from "./src/local";
+export {default as matcher} from "./src/matcher";
+export {default as mouse} from "./src/mouse";
+export {default as namespace} from "./src/namespace";
+export {default as namespaces} from "./src/namespaces";
+export {default as select} from "./src/select";
+export {default as selectAll} from "./src/selectAll";
+export {default as selection} from "./src/selection/index";
+export {default as selector} from "./src/selector";
+export {default as selectorAll} from "./src/selectorAll";
+export {default as touch} from "./src/touch";
+export {default as touches} from "./src/touches";
+export {default as window} from "./src/window";
+export {event, customEvent} from "./src/selection/on";
